Add port select in mission land control
