+++ frontend/src/features/SideWindow/MissionForm/ActionForm/LandControlForm.tsx
@@ -12,6 +12,7 @@ import { useMemo } from 'react'
 
 import { ControlQualityField } from './shared/ControlQualityField'
 import { FormikMultiInfractionPicker } from './shared/FormikMultiInfractionPicker'
+import { FormikPortSelect } from './shared/FormikPortSelect'
 import { GearsField } from './shared/GearsField'
 import { LicencesAndLogbookField } from './shared/LicencesAndLogbookField'
 import { SpeciesField } from './shared/SpeciesField'
@@ -61,6 +62,8 @@ export function LandControlForm({ initialValues, onChange }: LandControlFormProp
             name="actionDatetimeUtc"
             withTime
           />
+
+          <FormikPortSelect />
 
           <LicencesAndLogbookField />
