+++ frontend/src/features/SideWindow/MissionForm/MainForm/FormikLocationPicker.tsx
@@ -1,12 +1,11 @@
-import { Accent, Button, Fieldset, Icon, IconButton, Label, Select, type Option } from '@mtes-mct/monitor-ui'
+import { Accent, Button, Fieldset, Icon, IconButton, Label } from '@mtes-mct/monitor-ui'
 import { useFormikContext } from 'formik'
 import { boundingExtent } from 'ol/extent'
 import { transformExtent } from 'ol/proj'
 import { remove } from 'ramda'
-import { useCallback, useEffect, useMemo, useState } from 'react'
+import { useCallback, useEffect, useMemo } from 'react'
 import styled from 'styled-components'
 
-import { useGetPortsQuery } from '../../../../api/port'
 import {
   InteractionListener,
   OPENLAYERS_PROJECTION,
@@ -17,34 +16,15 @@ import { fitToExtent } from '../../../../domain/shared_slices/Map'
 import { addMissionZone } from '../../../../domain/use_cases/missions/addMissionZone'
 import { useListenForDrawedGeometry } from '../../../../hooks/useListenForDrawing'
 import { useMainAppDispatch } from '../../../../hooks/useMainAppDispatch'
-import { FrontendError } from '../../../../libs/FrontendError'
-import { useNewWindow } from '../../../../ui/NewWindow'
 
-import type { Port } from '../../../../domain/types/port'
 import type { MissionFormValues } from '../types'
 import type { Coordinate } from 'ol/coordinate'
 
 export function FormikLocationPicker() {
   const { setFieldValue, values } = useFormikContext<MissionFormValues>()
 
-  const { newWindowContainerRef } = useNewWindow()
-
-  const [selectedDummyPort, setSelectedDummyPort] = useState<Port.Port | undefined>(undefined)
-
-  const getPortsApiQuery = useGetPortsQuery()
   const { geometry } = useListenForDrawedGeometry(InteractionListener.MISSION_ZONE)
   const dispatch = useMainAppDispatch()
-
-  const portsAsOptions: Option[] = useMemo(() => {
-    if (!getPortsApiQuery.data) {
-      return []
-    }
-
-    return getPortsApiQuery.data.map(({ locode, name }) => ({
-      label: `${name} (${locode})`,
-      value: locode
-    }))
-  }, [getPortsApiQuery.data])
 
   const polygons = useMemo(() => {
     if (!values.geom) {
@@ -87,38 +67,6 @@ export function FormikLocationPicker() {
     [dispatch]
   )
 
-  const handlePortChange = useCallback(
-    (nextPortLocode: string | undefined) => {
-      if (!getPortsApiQuery.data) {
-        return
-      }
-
-      if (!nextPortLocode) {
-        // setFieldValue('portLocode', undefined)
-        // setFieldValue('portName', undefined)
-
-        setSelectedDummyPort(undefined)
-
-        return
-      }
-
-      const port = getPortsApiQuery.data.find(({ locode }) => locode === nextPortLocode)
-      if (!port) {
-        throw new FrontendError('`port` is undefined')
-      }
-
-      // TODO We'll have to replace that with the port geometry.
-      setFieldValue('geom', undefined)
-      // setFieldValue('portLocode', port.locode)
-      // setFieldValue('portLocode', port.locode)
-
-      setSelectedDummyPort(port)
-    },
-
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    [getPortsApiQuery.data]
-  )
-
   useEffect(
     () => {
       if (geometry?.type === OpenLayersGeometryType.MULTIPOLYGON) {
@@ -135,7 +83,7 @@ export function FormikLocationPicker() {
       <Label>Localisations</Label>
 
       <div>
-        <Button accent={Accent.SECONDARY} disabled={!!selectedDummyPort} Icon={Icon.Plus} isFullWidth onClick={addZone}>
+        <Button accent={Accent.SECONDARY} Icon={Icon.Plus} isFullWidth onClick={addZone}>
           Ajouter une zone de mission
         </Button>
 
@@ -163,23 +111,6 @@ export function FormikLocationPicker() {
             </Row>
           ))}
         </>
-      </div>
-
-      <div>
-        <StyledSelect
-          baseContainer={newWindowContainerRef.current}
-          isLabelHidden
-          isLight
-          label="Lieu du contrôle"
-          name="port"
-          onChange={handlePortChange}
-          options={portsAsOptions}
-          placeholder="ou sélectionnez un port"
-          searchable
-          // value={values.portLocode}
-          value={selectedDummyPort?.locode}
-          virtualized
-        />
       </div>
     </Fieldset>
   )
@@ -215,8 +146,4 @@ const Link = styled.a`
     line-height: 1;
     margin: -2px 0 0 8px;
   }
-`
-
-const StyledSelect = styled(Select<string>)`
-  margin-top: 12px;
 `
